Refactoring activation extraction for all RNN layers
